delete PipelineDefaultGroupCollapsed logic
